# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Template Todo App API",
    "version": "0.1.0",
    "description": "Todo アプリケーション API のリクエスト・レスポンス仕様。zod スキーマから自動生成。"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {},
    "schemas": {
...
      "taskResponseSchema": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "task": {
            "$ref": "#/components/schemas/taskSchema"
          }
        },
        "required": [
          "success",
          "task"
        ]
      },
...
      "tasksListResponseSchema": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/taskSchema"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/paginationSchema"
          }
        },
        "required": [
          "success",
          "tasks",
          "pagination"
        ]
      },
      "createTaskSchema": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "タスクのタイトル（1〜100文字）"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "タスクの説明（最大1000文字、空文字可）"
          },
          "accountId": {
            "type": "string",
            "minLength": 1,
            "description": "所属するアカウントのID"
          }
        },
        "required": [
          "title",
          "description",
          "accountId"
        ]
      },
...
      "taskStatusSchema": {
        "type": "string",
        "description": "タスクの状態"
      },
      "taskSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "タスクID"
          },
          "accountId": {
            "type": "string",
            "description": "所属アカウントID"
          },
          "title": {
            "type": "string",
            "description": "タイトル"
          },
          "description": {
            "type": "string",
            "description": "説明"
          },
          "status": {
            "$ref": "#/components/schemas/taskStatusSchema"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "作成日時"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新日時"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "論理削除日時（未削除なら null）"
          }
        },
        "required": [
          "id",
          "accountId",
          "title",
          "description",
          "status",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "paginationSchema": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "現在のページ"
          },
          "limit": {
            "type": "integer",
            "description": "1ページあたりの件数"
          },
          "total": {
            "type": "integer",
            "description": "総件数"
          },
          "totalPages": {
            "type": "integer",
            "description": "総ページ数"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request - 入力バリデーションに失敗、または不正な JSON ボディ",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "入力値が不正です"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                },
                "issues": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  },
                  "description": "VALIDATION_ERROR のときのみ存在"
                }
              },
              "required": [
                "success",
                "error",
                "code"
              ]
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error - サーバー内部エラー",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "サーバー内部エラー"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                },
                "issues": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  },
                  "description": "VALIDATION_ERROR のときのみ存在"
                }
              },
              "required": [
                "success",
                "error",
                "code"
              ]
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "operationId": "get-tasks",
        "summary": "タスク一覧を取得",
        "description": "指定アカウントのタスクをページング付きで取得する",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "accountId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "対象アカウントのID"
            },
            "description": "対象アカウントのID"
          },
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "default": 1,
              "description": "ページ番号（1始まり）",
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "description": "ページ番号（1始まり）"
          },
          {
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "default": 8,
              "description": "1ページあたりの件数（1〜100）",
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "1ページあたりの件数（1〜100）"
          }
        ],
        "responses": {
          "200": {
            "description": "タスク一覧とページング情報",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tasksListResponseSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-tasks",
        "summary": "タスクを作成",
        "description": "タスクを新規作成する。ステータスは TODO で初期化される。",
        "tags": [
          "Tasks"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createTaskSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "作成されたタスク",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/taskResponseSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}